--- README.md
@@ -1,28 +1,0 @@
-# streamlit-sandbox
-Sandbox for testing features and use cases for streamlit library
-
-#### Run the following command for starting application
-
-Create virtual environment:
-
- `python3 -m venv env`
- 
- Activate Virtual environment:
- 
- `source env/bin/activate`
- 
- You can deactivate the virtual environment as well:
- 
- `deactivate`
- 
-Install necessary libraries:
-
- - `pip install streamlit`
-
- - `pip install spacy`
-
- - `python -m spacy download en_core_web_sm`
-
-Run the application with the following command:
-
- `streamlit run app.py`
